requirements.txt
pybit
pandas
numpy
fastapi
uvicorn[standard]
jinja2
python-dotenv
